+++ __tests__/helpers/__mocks__/virtual-mermaid-config.ts
@@ -2,4 +2,6 @@ export default {
   securityLevel: 'loose',
   startOnLoad: false,
   externalDiagrams: [],
+  download: true,
+  downloadPng: true,
 };